requirements.txt
streamlit
pymongo
python-dotenv
pandas
altair
Flask
Flask-Cors
gunicorn